+++ tests/PHPStan/Testing/data/assert-certainty-case-insensitive.php
