+++ tests/PHPStan/Testing/data/assert-certainty-missing-namespace.php
