+++ tests/PHPStan/Testing/data/assert-certainty-wrong-namespace.php
