+++ tests/PHPStan/Testing/data/assert-native-type-case-insensitive.php
