+++ tests/PHPStan/Testing/data/assert-native-type-missing-namespace.php
