+++ tests/PHPStan/Testing/data/assert-native-type-wrong-namespace.php
