+++ tests/PHPStan/Testing/data/assert-type-case-insensitive.php
