+++ tests/PHPStan/Testing/data/assert-type-missing-namespace.php
