+++ tests/PHPStan/Testing/data/assert-type-wrong-namespace.php
